pyproject.toml
[project]
name = "langgraph-01"
version = "0.1.0"
description = "LangGraph Practice Environment"
requires-python = ">=3.12"

dependencies = [
    "langchain==0.3.11",
    "langchain-community==0.3.11",
    "langchain-openai==0.2.3",
    "openai>=1.56",
    "langgraph==0.3.18",
    "langgraph-checkpoint-sqlite==2.0.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "gradio>=5.21.0",
    "playwright>=1.56.0",
]
